refactor(kitchenpos-product): MenuGroup 패키지 경로 수정
