+++ src/main/java/kitchenpos/menus/application/MenuGroupService.java
@@ -1,10 +1,9 @@
 package kitchenpos.menus.application;
 
 import java.util.List;
-import java.util.Objects;
-import java.util.UUID;
-import kitchenpos.menus.domain.MenuGroup;
 import kitchenpos.menus.domain.MenuGroupRepository;
+import kitchenpos.menus.domain.tobe.MenuGroup;
+import kitchenpos.menus.domain.tobe.MenuGroupName;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,15 +17,8 @@ public class MenuGroupService {
     }
 
     @Transactional
-    public MenuGroup create(final MenuGroup request) {
-        final String name = request.getName();
-        if (Objects.isNull(name) || name.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-        final MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(UUID.randomUUID());
-        menuGroup.setName(name);
-        return menuGroupRepository.save(menuGroup);
+    public MenuGroup create(final MenuGroupName request) {
+        return menuGroupRepository.save(new MenuGroup(request));
     }
 
     @Transactional(readOnly = true)
