+++ src/main/java/kitchenpos/menus/domain/Menu.java
@@ -13,9 +13,10 @@ import jakarta.persistence.Transient;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
+import kitchenpos.menus.domain.tobe.MenuGroup;
 
 @Table(name = "menu")
-@Entity
+//@Entity
 public class Menu {
 
     @Column(name = "id", columnDefinition = "binary(16)")
