+++ src/main/java/kitchenpos/menus/domain/MenuGroupRepository.java
@@ -3,6 +3,7 @@ package kitchenpos.menus.domain;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.menus.domain.tobe.MenuGroup;
 
 public interface MenuGroupRepository {
